utils/yaml.py: create and populate
OSBS-8786

Move read_yaml_from_file_path and dependent functions out of repo_utils
and into utils/yaml.py
